requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.2
Werkzeug==2.3.7
python-dotenv==1.0.0
ultralytics==8.0.196
torch==2.2.0
torchvision==0.17.0
numpy==1.24.3
Pillow==10.0.0
joblib
